requirements.txt
torch>=2.0
torchvision>=0.15
opencv-python
numpy
matplotlib
tqdm


# for app-API
fastapi
uvicorn
torch
opencv-python-headless
torchvision
numpy
Pillow
matplotlib
python-multipart

# streamlit APP
streamlit
torch
opencv-python-headless
torchvision
numpy
Pillow
matplotlib
